chore: restructure collector profile header files
