+++ src/Apps/CollectorProfile/Components/CollectorProfileHeader/Components/CollectorProfileHeaderAvatar.tsx
