+++ src/Apps/CollectorProfile/Components/CollectorProfileHeader/Components/CollectorProfileHeaderInfo.tsx
