+++ src/Apps/CollectorProfile/Components/CollectorProfileHeader/Components/__tests__/CollectorProfileHeaderAvatar.jest.tsx
@@ -1,5 +1,5 @@
 import { screen } from "@testing-library/react"
-import { CollectorProfileHeaderAvatarFragmentContainer } from "Apps/CollectorProfile/Components/CollectorProfileHeaderAvatar"
+import { CollectorProfileHeaderAvatarFragmentContainer } from "Apps/CollectorProfile/Components/CollectorProfileHeader/Components/CollectorProfileHeaderAvatar"
 import { MockBoot } from "DevTools"
 import { setupTestWrapperTL } from "DevTools/setupTestWrapper"
 import { graphql } from "react-relay"
