+++ src/Apps/CollectorProfile/Components/CollectorProfileHeader/Components/__tests__/CollectorProfileHeaderInfo.jest.tsx
@@ -1,5 +1,5 @@
 import { screen } from "@testing-library/react"
-import { CollectorProfileHeaderInfoFragmentContainer } from "Apps/CollectorProfile/Components/CollectorProfileHeaderInfo"
+import { CollectorProfileHeaderInfoFragmentContainer } from "Apps/CollectorProfile/Components/CollectorProfileHeader/Components/CollectorProfileHeaderInfo"
 import { MockBoot } from "DevTools"
 import { setupTestWrapperTL } from "DevTools/setupTestWrapper"
 import { graphql } from "react-relay"
