+++ src/Apps/CollectorProfile/Components/CollectorProfileHeader/__tests__/CollectorProfileHeader.jest.tsx
@@ -1,6 +1,6 @@
 import { Breakpoint } from "@artsy/palette"
 import { screen } from "@testing-library/react"
-import { CollectorProfileHeaderFragmentContainer } from "Apps/CollectorProfile/Components/CollectorProfileHeader"
+import { CollectorProfileHeaderFragmentContainer } from "Apps/CollectorProfile/Components/CollectorProfileHeader/CollectorProfileHeader"
 import { MockBoot } from "DevTools"
 import { setupTestWrapperTL } from "DevTools/setupTestWrapper"
 import { graphql } from "react-relay"
